executor: implements nested loop join.
