+++ executor/executor.go
@@ -42,8 +42,6 @@ var (
 	_ Executor = &DistinctExec{}
 	_ Executor = &DummyScanExec{}
 	_ Executor = &ExistsExec{}
-	_ Executor = &HashAggExec{}
-	_ Executor = &HashJoinExec{}
 	_ Executor = &HashSemiJoinExec{}
 	_ Executor = &LimitExec{}
 	_ Executor = &MaxOneRowExec{}
